реализует сортировку фильмов
